femmenourish/requirements.txt
# Core Framework
streamlit==1.31.0
python-dotenv==1.0.1

# AI/ML APIs
google-generativeai==0.3.2

# HTTP Requests
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
pyyaml==6.0.1

# Image Processing (for ultrasound analysis)
Pillow==10.2.0
opencv-python==4.9.0.80

# Visualization
plotly==5.18.0

# PDF Generation
reportlab==4.0.9

# Streamlit Extensions
streamlit-option-menu==0.3.12

# Utilities
python-dateutil==2.8.2
